Add evolutionary algorithm and refactor routes and cost function
Add evolutionary algorithm.  Uses roulette selection for recombination.
Two parents are chosen, and they produce one child.  This is done
enough times to double the population.  Children are then potentially
mutated by swapping two nodes.  Population control is done via roulette
selection, removing half the individuals to restore the original
population size.

Because of the recombination, the routes and cost function have been
refactored to not include the loop in the route but, instead add it
when calculating the cost.